% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dj_connect.R
\name{dj_connect}
\alias{dj_connect}
\title{Connect to datajoint database}
\usage{
dj_connect(conda = NULL, venv = NULL)
}
\arguments{
\item{conda}{conda environment to use. May only specify one of conda or venv}

\item{venv}{virtual environment to use. May only specify one of conda or venv}
}
\description{
Connect to datajoint database
}
\examples{
dj_connect()

}
