% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dj_import.R
\name{dj_import}
\alias{dj_import}
\title{Import datajoint-python module}
\usage{
dj_import(module = ".", schema = "*")
}
\arguments{
\item{module}{the name of the python module containing your schemas. Default = ".", the local working directory}

\item{schema}{the names of the schema(s) to import. Default = "*" (imports all modules)}
}
\description{
Imports module (and schema) from which to fetch data
}
