% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_from_log.R
\name{get_table_from_log}
\alias{get_table_from_log}
\title{Get datajoint table}
\usage{
get_table_from_log(sch, tbl, log, module = "schema", callback = NULL,
  add_vars = list(), ...)
}
\arguments{
\item{sch}{the names of the schema used to fetch; string}

\item{tbl}{the table to fetch; string}

\item{log}{a data frame with column names as key and rows as value for restrictions}

\item{module}{the name of the python module containing your schemas}

\item{callback}{a callback function to process data returned from the query}

\item{add_vars}{a list of addititional variables to add to the data, optional}

\item{...}{additional arguments passed to callback}
}
\value{
Returns a list. if callback is specified, contents of the list are in the form of the callback's return. otherwise, returns a list of lists.
}
\description{
Fetches data from a datajoint table, using restrictions from a log (a data frame)
}
