% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config.R
\name{set_config}
\alias{set_config}
\title{Set datajoint configuration}
\usage{
set_config(host = NULL, user = NULL, pw = NULL, save_global = T)
}
\arguments{
\item{host}{ip address of host database}

\item{user}{user name}

\item{pw}{password}

\item{save_global}{boolean; if TRUE, save configuration globally. If FALSE, save configuration locally}
}
\value{
None
}
\description{
Sets the datajoint.config with desired hostname, username, and password
and saves the edited datajoint.config either locally or globally.
}
