% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_python.R
\name{set_python}
\alias{set_python}
\title{Set python version}
\usage{
set_python(conda = NULL, venv = NULL)
}
\arguments{
\item{conda}{name of conda environment, one of conda or venv must be specified}

\item{venv}{name of virtual environment, one of conda or venv must be specified}
}
\value{
None
}
\description{
Set python version
}
\examples{
set_python(conda="dj")

}
